storage/tests: Move kvstore_fixture into a header
